Create ImageReadPlot
Using Python's Pillow library, read in the image of a circle drawn by a contestant. Convert the image to gray scale, and plot a set of discrete points to use later in circular regression. The circles were drawn on an iPad. This preserves the original form of the drawn circles.